Add support for watches and validator
Have a Refreshable implement:
 - in Clojure: IRef,
 - and in ClojureScript: IWatchable and IDeref

to allow for the `set-validator!` and adding, removing, and triggering
of watches of a Refreshable.

In ClojureScript, have the IDeref give an Atom like behavior, by
using the current value as returned from `core.async/poll!`, so that
if the value has not been initialized, it treats the value as `nil`,
since there is no value set on the promise-chan before initialization
takes effect.